Adding more playbooks

+++ simple-node/scripts/stop.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 ID=`ps -ef | grep npm | grep start | tr -s ' ' | cut -d' ' -f2`
-kill -9 $ID
+if [[ "x$ID" -eq "x" ]]
+then
+  echo "Application is not running"
+else
+  kill -9 $ID
+  echo "Killing application running with process ID $ID"
+fi
